Improving prowler
added a new behavior type, "prowl". loses track of where the player is, and thus begins quickly searching the near vicinity of where it detected them last. Code is messy and far from perfect: cases to detect player are locked inside of scripts that are only called every so often, resulting in a bug where the player is sometimes not detected when they should be.